+++ crates/rt/src/storage/records.rs
@@ -104,7 +104,7 @@ pub struct ChannelRecord {
     #[serde(default, skip_serializing_if = "is_false")]
     pub is_closed: bool,
     /// Number of messages written to the channel.
-    pub received_messages: usize,
+    pub received_messages: u64,
 }
 
 impl ChannelRecord {
@@ -148,7 +148,7 @@ pub struct WorkflowRecord<T = WorkflowState> {
     /// Name of the workflow in the module.
     pub name_in_module: String,
     /// Number of the workflow executions.
-    pub execution_count: usize,
+    pub execution_count: u64,
     /// Current state of the workflow.
     pub state: T,
 }
@@ -304,7 +304,7 @@ pub struct ErroneousMessageRef {
     /// ID of the channel the message was received from.
     pub channel_id: ChannelId,
     /// 0-based index of the message in the channel.
-    pub index: usize,
+    pub index: u64,
 }
 
 /// Workflow selection criteria used in [`insert_waker_for_matching_workflows()`].
